Refactor reindex helpers in an importable way
